refactor: 패키지 구조에 따른 테스트 분리
