+++ src/test/java/cart/controller/AdminControllerTest.java
@@ -1,21 +1,8 @@
 package cart.controller;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.doNothing;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 import cart.dto.ProductDto;
 import cart.service.ProductManagementService;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.ArrayList;
-import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -24,6 +11,13 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.doNothing;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
@@ -40,24 +34,11 @@ public class AdminControllerTest {
     private ProductManagementService productManagementService;
 
     @Test
-    void 어드민_페이지를_조회한다() throws Exception {
-        List<ProductDto> products = new ArrayList<>();
-        given(productManagementService.findAllProduct())
-                .willReturn(products);
-
-        mockMvc.perform(get("/admin"))
-                .andExpect(model().attributeExists("products"))
-                .andExpect(model().attribute("products", products))
-                .andExpect(status().isOk());
-    }
-
-    @Test
     void 상품을_추가한다() throws Exception {
         final ProductDto productDto = new ProductDto("하디", "imageUrl", 100000);
         final String requestBody = objectMapper.writeValueAsString(productDto);
-        doNothing()
-                .when(productManagementService)
-                .addProduct(any());
+        given(productManagementService.addProduct(any()))
+                .willReturn(anyLong());
 
         mockMvc.perform(post("/admin/products")
                         .content(requestBody)
@@ -74,10 +55,10 @@ public class AdminControllerTest {
                 .when(productManagementService)
                 .updateProduct(any());
 
-        mockMvc.perform(patch("/admin/products/{product_id}", id)
+        mockMvc.perform(put("/admin/products/{product_id}", id)
                         .content(requestBody)
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andExpect(status().isCreated());
+                .andExpect(status().isNoContent());
     }
 
     @Test
@@ -89,6 +70,6 @@ public class AdminControllerTest {
                 .deleteProduct(anyLong());
 
         mockMvc.perform(delete("/admin/products/{product_id}", id))
-                .andExpect(status().isOk());
+                .andExpect(status().isNoContent());
     }
 }
